GLV V 12bit displaying; attempting CAN recv detection

+++ Dashboard.cydsn/frucd_display.h
@@ -29,11 +29,11 @@ void initDashTemplate();
  * BEGIN: functions to show data on dash
  * - color decision logic is in here
  */
-void disp_SOC(int data);
-void disp_max_pack_temp(int data);
+void disp_SOC(uint8 data);
+void disp_max_pack_temp(uint8 data);
 void disp_state(uint8 state); // TODO
-void disp_glv_v(int data); // TODO
-void disp_mc_temp(int data);
-void disp_motor_temp(int data);
+void disp_glv_v(uint16 data); // TODO
+void disp_mc_temp(uint8 data);
+void disp_motor_temp(uint8 data);
 
 /* [] END OF FILE */
